Update main layout; update banner
