+++ src/components/Carousel/Carousel.css
@@ -1,0 +1,45 @@
+.slide-item {
+    object-fit: cover;
+}
+
+.home-banner {
+    margin-top: 64px;
+}
+
+.back-btn {
+    position: absolute;
+    left: 0;
+    bottom: 50%;
+    border: none;
+    background: none;
+    margin-left: 3.5rem;
+}
+
+.next-btn {
+    position: absolute;
+    right: 0;
+    bottom: 50%;
+    border: none;
+    background: none;
+    margin-right: 3.5rem;
+}
+
+.dot-group-btns {
+    position: absolute;
+    bottom: 3rem;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    display: flex;
+    column-gap: 8px;
+}
+
+.dot-group-btns > * {
+    border-radius: 50%;
+    padding: 0.3rem;
+    /* border: grey 3px solid; */
+}
+
+.dot-group-btns > button:not(.carousel__dot--selected) {
+    background-color: #2C2F35 !important;
+    /* border: grey 3px solid; */
+}
