--- src/components/Header/Header.css
@@ -1,93 +1,0 @@
-.banner {
-    height: 350px;
-    margin-top: 64px;
-    background-size: cover;
-    background-position: "center";
-}
-
-.home-banner {
-    margin-top: 64px;
-}
-
-.slide-item {
-    object-fit: cover;
-}
-
-.back-btn {
-    position: absolute;
-    left: 0;
-    bottom: 50%;
-    border: none;
-    background: none;
-    margin-left: 3.5rem;
-}
-
-.next-btn {
-    position: absolute;
-    right: 0;
-    bottom: 50%;
-    border: none;
-    background: none;
-    margin-right: 3.5rem;
-}
-
-.dot-group-btns {
-    position: absolute;
-    bottom: 3rem;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    display: flex;
-    column-gap: 8px;
-}
-
-.dot-group-btns > * {
-    border-radius: 50%;
-    padding: 0.3rem;
-    /* border: grey 3px solid; */
-}
-
-.dot-group-btns > button:not(.carousel__dot--selected) {
-    background-color: #2C2F35 !important;
-    /* border: grey 3px solid; */
-}
-
-.banner-title {
-    position: relative;
-    background: #6F0000;
-    opacity: 80%;
-    display: inline-block;
-    color: white;
-    padding: .24em .3em .25em;
-    padding: 1rem;  
-    z-index: 1000;
-    width: 35%;
-    height: inherit;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    text-transform: uppercase;
-    font-size: 2.5em;
-    font-weight: 700;
-}
-
-.banner-title:after {
-    position: absolute;
-    z-index: -1;
-    content: "";
-    left: 13%;
-    top: 0;
-    height: 100%;
-    width: 100%;
-    background-color: inherit;
-    -webkit-transform: skewX(20deg);
-    -moz-transform: skewX(20deg);
-    -ms-transform: skewX(20deg);
-    transform: skewX(20deg);
-}
-
-#banner-title {
-    font-size: 12px !important;
-    font-weight: 700;
-    text-transform: uppercase;
-    margin: 0 !important;
-}
