+++ src/components/Navbar/Navbar.css
@@ -1,7 +1,7 @@
 .nav-style {
     width: 100%;
     background-color: #2C2F35;
-    position: 'fixed';
+    position: 'sticky';
     z-index: 1000;
     top: 0;
 }
