+++ src/components/Wrappers/MainContent/MainContent.js
@@ -3,13 +3,13 @@ import "./MainContent.css";
 
 const MainContent = ({ title, children }) => {
   return (
-    <div>
-    <div>
-        <h1 className='mc-title' id='mc-title'>{title}</h1>
-    </div>
-    <div style={{ backgroundColor: "#FFF"}}>
-        {children}
-    </div>
+    <div style={{ display: "flex", flexDirection: "column"}} >
+      <div>
+          <h1 className='mc-title' id='mc-title'>{title}</h1>
+      </div>
+      <div style={{ backgroundColor: "#FFF"}}>
+          {children}
+      </div>
     </div>
   )
 }
