+++ src/router/AppRouter.js
@@ -1,41 +1,25 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
-import Header from '../components/Header/Header';
 import Navbar from '../components/Navbar/Navbar';
 import Footer from '../components/Footer/Footer';
 import Home from '../pages/Home/Home';
-import Area from '../pages/Area/Area';
 import Competitions from '../pages/Competitions/Competitions';
 import Teams from '../pages/Teams/Teams';
 import Match from '../pages/Match/Match';
 import News from '../pages/News/News';
 import Auth from '../pages/Auth/Auth';
-import CommentsCol from '../components/Layout/CommentsCol/CommentsCol';
-import { Container, Row } from 'react-bootstrap';
-
+import Layout from '../components/Layout/Main/Main';
 import ProtectedRoutes from '../components/ProtectedRoutes';
 
 const token = localStorage.getItem("token");
-
-const Layout = ({children}) => {
-  return(
-    <Container className='layout'>
-      <Row>
-        <CommentsCol />
-        {children}
-      </Row>
-    </Container>
-  )
-};
-
+const path = window.location.pathname;
 
 const AppRouter = () => {
   
   return (
       <Router>
       <Navbar token={token} />
-      <Header />
-      <Layout>
+      <Layout path={path}>
         <Routes>
           <Route path="/"  element={<Home /> } />
           <Route path="/teams"  element={<Teams/>} />
